UPDATE split roomEntity

+++ src/Hotel/reserveBundle/Controller/HotelRoomController.php
@@ -6,14 +6,14 @@ use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Hotel\reserveBundle\Entity\roomEntity;
 use Hotel\reserveBundle\Form\roomEntityType;
 
-class roomEntityController extends Controller
+class HotelRoomController extends Controller
 {
     public function indexAction($hotel_id)
     {
         $em = $this->getDoctrine()->getManager();
         $hotel = $em->getRepository('HotelreserveBundle:hotelEntity')->find($hotel_id);
 
-        if(!$hotel)
+        if(!$hotel || !$hotel->getHotelActive())
             throw $this->createNotFoundException('Unable to find hotel entity.');
 
         $entities = $hotel->getRoomEntities();
@@ -25,7 +25,7 @@ class roomEntityController extends Controller
             10
         );
 
-        return $this->render('HotelreserveBundle:roomEntity:index.html.twig', array(
+        return $this->render('HotelreserveBundle:hotel_room:index.html.twig', array(
             'entities' => $pagination,
             'hotel_id' => $hotel_id
         ));
@@ -33,6 +33,12 @@ class roomEntityController extends Controller
 
     public function newAction(Request $request,$hotel_id)
     {
+        $em = $this->getDoctrine()->getManager();
+
+        $hotel = $em->getRepository('HotelreserveBundle:hotelEntity')->find($hotel_id);
+        if(!$hotel || !$hotel->getHotelActive())
+            throw $this->createNotFoundException('Unable to find hotel entity.');
+
         $entity = new roomEntity();
         $form = $this->createForm(new roomEntityType(), $entity);
 
@@ -41,17 +47,15 @@ class roomEntityController extends Controller
             $form->handleRequest($request);
 
             if ($form->isValid()) {
-                $em = $this->getDoctrine()->getManager();
-                $hotel = $em->getRepository('HotelreserveBundle:hotelEntity')->find($hotel_id);
                 $entity->setHotelEntity($hotel);
                 $em->persist($entity);
                 $em->flush();
 
-                return $this->redirect($this->generateUrl('roomentity', array('id' => $entity->getId(),'hotel_id'=>$hotel_id)));
+                return $this->redirect($this->generateUrl('hotel_room_index', array('hotel_id'=>$hotel_id)));
             }
         }
 
-        return $this->render('HotelreserveBundle:roomEntity:new.html.twig', array(
+        return $this->render('HotelreserveBundle:hotel_room:new.html.twig', array(
             'entity' => $entity,
             'form'   => $form->createView(),
             'hotel_id' => $hotel_id
@@ -62,11 +66,13 @@ class roomEntityController extends Controller
     {
         $em = $this->getDoctrine()->getManager();
 
-        $entity = $em->getRepository('HotelreserveBundle:roomEntity')->find($id);
+        $hotel = $em->getRepository('HotelreserveBundle:hotelEntity')->find($hotel_id);
+        if(!$hotel || !$hotel->getHotelActive())
+            throw $this->createNotFoundException('Unable to find hotel entity.');
 
-        if (!$entity) {
+        $entity = $em->getRepository('HotelreserveBundle:roomEntity')->find($id);
+        if (!$entity || $entity->getHotelEntity() != $hotel)
             throw $this->createNotFoundException('Unable to find roomEntity entity.');
-        }
 
         $editForm = $this->createForm(new roomEntityType(), $entity);
 
@@ -77,11 +83,11 @@ class roomEntityController extends Controller
             if ($editForm->isValid()) {
                 $em->flush();
 
-                return $this->redirect($this->generateUrl('roomentity', array('id' => $id,'hotel_id'=>$hotel_id)));
+                return $this->redirect($this->generateUrl('hotel_room_index', array('hotel_id'=>$hotel_id)));
             }
         }
 
-        return $this->render('HotelreserveBundle:roomEntity:edit.html.twig', array(
+        return $this->render('HotelreserveBundle:hotel_room:edit.html.twig', array(
             'entity'      => $entity,
             'edit_form'   => $editForm->createView(),
             'hotel_id'    => $hotel_id
@@ -91,11 +97,14 @@ class roomEntityController extends Controller
     public function deleteAction($id,$hotel_id)
     {
         $em = $this->getDoctrine()->getManager();
-        $entity = $em->getRepository('HotelreserveBundle:roomEntity')->find($id);
 
-        if (!$entity) {
+        $hotel = $em->getRepository('HotelreserveBundle:hotelEntity')->find($hotel_id);
+        if(!$hotel || !$hotel->getHotelActive())
+            throw $this->createNotFoundException('Unable to find hotel entity.');
+
+        $entity = $em->getRepository('HotelreserveBundle:roomEntity')->find($id);
+        if (!$entity)
             throw $this->createNotFoundException('Unable to find roomEntity entity.');
-        }
 
         foreach($entity->getBlankEntities() as $blank)
             $em->remove($blank);
@@ -103,6 +112,6 @@ class roomEntityController extends Controller
         $em->remove($entity);
         $em->flush();
 
-        return $this->redirect($this->generateUrl('roomentity',array('hotel_id'=>$hotel_id)));
+        return $this->redirect($this->generateUrl('hotel_room_index',array('hotel_id'=>$hotel_id)));
     }
 }
